Converting bytestring tag name to utf8
MySQLdb returns bytestring rather than unicode if column is set to collate utf8_bin. If column contains utf8 characters such as ñ or á, and db column has collate utf8_bin, a DjangoUnicodeDecodeError when django template tries to render the field using force_unicode(value). 

This change is to convert the byte string to utf8 unicode (if it is not already unicode) before being returned.